fix(correcao paginacao lista da execucao orcamentaria painel estrategico)

+++ backend/src/gestao-projetos/painel-estrategico/painel-estrategico.service.ts
@@ -654,7 +654,7 @@ export class PainelEstrategicoService {
         body: AnyPageTokenJwtBody;
     }> {
         const quantidade_rows = await this.prisma.$queryRawUnsafe(`select
-                                                                       count(*)
+                                                                       count(*)::int
                                                                    from
                                                                        (select
                                                                             pr.id
